fix(build): move server utility functions to helper file
