+++ package.json
@@ -96,5 +96,6 @@
     "supabase": "^1.162.4",
     "typescript": "^5.4.5",
     "vitest": "^1.5.0"
-  }
+  },
+  "packageManager": "pnpm@9.0.4"
 }
